requirements.txt
requests
pytest
pytest-mock
fastapi[all]
uvicorn
python-dotenv